+++ .gitignore
@@ -1,3 +1,3 @@
 build
-src/libpds.o
+src/*.o
 .vscode
